pyproject.toml
[project]
name = "sw-metadata-bot"
version = "0.1.0"
description = "RSMetaCheck bot for pushing issues with existing repository metadata"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Tom François", email = "tom.francois@lapp.in2p3.fr"}
]
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
